feat(signup): - refactored signup process on home and signup components - moved all api calls from actions to new file DialApi - added test files exclude to .codeclimate.yml [Finishes #152805872]
