+++ client/src/app/Helpers.js
@@ -2,7 +2,8 @@
  * defines a Helpers class
  */
 class Helpers {
-  static host = 'https://andela-events-manager.herokuapp.com/api/v1';
+  // static host = 'https://andela-events-manager.herokuapp.com/api/v1';
+  static host = 'http://localhost:7777/api/v1';
 
   static cloudinaryPreset = 'axgrmj0a';
